+++ stories/client-deliberation/DisagreeButton.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import DisagreeButton from '../../polis-UT-HAI/client-deliberation/src/components/DisagreeButton'
+import DisagreeButton from '../../codebases/UT-HAI/client-deliberation/src/components/DisagreeButton'
 import { action } from '@storybook/addon-actions'
 import { withDelibThemeUi } from '../../.storybook/decorators'
