+++ stories/client-deliberation/DoesNotExist.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import DoesNotExist from '../../polis-UT-HAI/client-deliberation/src/components/DoesNotExist'
+import DoesNotExist from '../../codebases/UT-HAI/client-deliberation/src/components/DoesNotExist'
 import { withDelibThemeUi } from '../../.storybook/decorators'
 
 export default {
