+++ stories/client-deliberation/Lock.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Lock from '../../polis-UT-HAI/client-deliberation/src/components/Lock'
+import Lock from '../../codebases/UT-HAI/client-deliberation/src/components/Lock'
 
 export default {
   title: 'Client-Deliberation/Lock',
