+++ stories/client-deliberation/TreeSummary.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import TreeSummary from '../../polis-UT-HAI/client-deliberation/src/components/TreeSummary'
+import TreeSummary from '../../codebases/UT-HAI/client-deliberation/src/components/TreeSummary'
 import { action } from '@storybook/addon-actions'
 import { withDelibThemeUi } from '../../.storybook/decorators'
