+++ stories/client-deliberation/TutorialBackground.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import TutorialBackground from '../../polis-UT-HAI/client-deliberation/src/components/TutorialBackground'
+import TutorialBackground from '../../codebases/UT-HAI/client-deliberation/src/components/TutorialBackground'
 import { withDelibThemeUi } from '../../.storybook/decorators'
 
 export default {
